Build fix.
svn:sync-xref-src-repo-rev: r113218

+++ src/VBox/Devices/Audio/DrvHostALSAAudio.cpp
@@ -1,4 +1,4 @@
-/* $Id: DrvHostALSAAudio.cpp 65565 2017-02-01 14:11:10Z andreas.loeffler@oracle.com $ */
+/* $Id: DrvHostALSAAudio.cpp 65566 2017-02-01 14:21:22Z andreas.loeffler@oracle.com $ */
 /** @file
  * VBox audio devices: ALSA audio driver.
  */
@@ -1182,7 +1182,7 @@ static DECLCALLBACK(int) drvHostALSAAudioStreamPlay(PPDMIHOSTAUDIO pInterface, P
             break;
         }
 
-        size_t cbToRead = RT_MIN(PDMAUDIOPCMPROPS_S2B(&pStreamALSA->Props, cAvail), cbBuf);
+        size_t cbToRead = RT_MIN((unsigned)PDMAUDIOPCMPROPS_S2B(&pStreamALSA->Props, cAvail), cbBuf);
         if (!cbToRead)
             break;
